Platforms are now visible in the editor (SUPER HACKY)
Resizeable platforms are now tiled based on their x scale. So if you set
the x scale to 4, the platform will be made up of 4 blocks. The editor
view just shows 1 block scaled to the width that the entire platform
would take up in-game (that one block sprite then gets scaled back to
normal at runtime).